# data/swagger.json
{"openapi":"3.0.0","paths":{"/api/v1":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/api/v1/user/create-user":{"post":{"operationId":"UserController_SignUp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreateDto"}}}},"responses":{"201":{"description":""}},"tags":["modules.user"]}},"/api/v1/user/login":{"post":{"operationId":"UserController_SignIn","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLoginDto"}}}},"responses":{"201":{"description":""}},"tags":["modules.user"]}},"/api/v1/tweet/create-tweet":{"post":{"operationId":"TweetController_CreateTweet","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTweetDto"}}}},"responses":{"201":{"description":""}},"tags":["modules.tweet"]}},"/api/v1/file/upload/video":{"post":{"operationId":"FileController_UploadFileVideo","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/FileSingleDto"}}}},"responses":{"201":{"description":""}},"tags":["modules.file"]}},"/api/v1/file/video-hls/{id}/master.m3u8":{"get":{"operationId":"FileController_serveM3u8","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["modules.file"]}},"/api/v1/file/video-hls/{id}/{v}/{segment}":{"get":{"operationId":"FileController_ServeSegmentController","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"v","required":true,"in":"path","schema":{"type":"string"}},{"name":"segment","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["modules.file"]}},"/api/v1/file/test":{"get":{"operationId":"FileController_test","parameters":[],"responses":{"200":{"description":""}},"tags":["modules.file"]}}},"info":{"title":"twitter, APIs Specification","description":"Section for describe whole APIs","version":"1.0","contact":{}},"tags":[],"servers":[{"url":"/"}],"components":{"securitySchemes":{"accessToken":{"scheme":"bearer","bearerFormat":"JWT","type":"http"},"refreshToken":{"scheme":"bearer","bearerFormat":"JWT","type":"http"},"google":{"scheme":"bearer","bearerFormat":"JWT","type":"http"},"apiKey":{"type":"apiKey","in":"header","name":"x-api-key"}},"schemas":{"UserCreateDto":{"type":"object","properties":{"email":{"type":"string","example":"Reymundo45@hotmail.com"},"password":{"type":"string","description":"string password","example":"g7g48P85BN@@!123"},"username":{"type":"string","example":"Katherine_Reilly87","nullable":true}},"required":["email","password"]},"UserLoginDto":{"type":"object","properties":{"email":{"type":"string","example":"Reymundo45@hotmail.com"},"password":{"type":"string","description":"string password","example":"g7g48P85BN@@!123"}},"required":["email","password"]},"CreateTweetDto":{"type":"object","properties":{"type":{"type":"number","enum":[0,1,2,3]},"audience":{"type":"number","enum":[0,1]},"content":{"type":"string","maxLength":100,"example":"My image is beautiful"},"parent_id":{"default":"","type":"array","items":{"type":"string"}},"hashtags":{"nullable":true,"default":"","example":["picture"],"type":"array","items":{"type":"string"}},"media":{"example":[{"type":"image","url":"./url"}],"type":"array","items":{"type":"string"}}},"required":["type","audience","content","media"]},"FileSingleDto":{"type":"object","properties":{"file":{"type":"string","format":"binary","description":"Single file"}},"required":["file"]}}}}